// frontend/src/pages/Todoist/types.ts
export interface TaskFormData {
  content: string;
  description: string;
  due_string: string;
  due_lang: string;
  priority: number;
  labels?: string[];
  project_id: string | null;
}

export interface TaskFilter {
  showCompleted: boolean;
}